+++ lib/eager_eye/detectors/loop_association.rb
@@ -4,6 +4,7 @@ module EagerEye
   module Detectors
     class LoopAssociation < Base
       ITERATION_METHODS = %i[each map collect select find find_all reject filter filter_map flat_map
+                             each_with_index each_with_object reduce inject
                              find_each find_in_batches in_batches].freeze
       PRELOAD_METHODS = %i[includes preload eager_load].freeze
       SINGLE_RECORD_METHODS = %i[find find_by find_by! first first! last last! take take! second third fourth fifth
@@ -42,7 +43,7 @@ module EagerEye
         traverse_ast(ast) do |node|
           next unless iteration_block?(node)
 
-          block_var = extract_block_variable(node)
+          block_var = extract_iteration_variable(node)
           next unless block_var
 
           block_body = node.children[2]